+++ composeApp/src/commonMain/composeResources/files/CHANGELOG.md
@@ -1,3 +1,14 @@
+Version: 1.5.0
+Date: 15-05-2026
+Fixed:
+	- Updater on windows now launches with correct path when applicable (.jar is not updatable)
+  - The update page now correctly displays the change description
+Added:
+  - Plugin now can define wich os they support (linux, windows, macos)
+  - Check to ensure the plugin is compatible with the app version
+  - Plugin signature
+  - Library used in the about section
+----------------------------------------------------------------------------------------------------
 Version: 1.4.1
 Date: 12-05-2026
 Added